+++ src/op/add_weighted_edge.rs
@@ -13,8 +13,13 @@
 //!
 //! assert_eq!(graph, vec![vec![(1, 2), (2, 1)], vec![(2, -3)], Vec::new()]);
 //! ```
+extern crate alloc;
 
 use {
+    alloc::collections::{
+        BTreeMap,
+        BTreeSet,
+    },
     core::hash::{
         BuildHasher,
         Hash,
@@ -116,6 +121,30 @@ where
     }
 }
 
+impl<W> AddWeightedEdge<W> for Vec<BTreeSet<(usize, W)>>
+where
+    W: Ord,
+{
+    /// # Panics
+    ///
+    /// Panics if `s` is not in the graph.
+    fn add_weighted_edge(&mut self, s: usize, t: usize, w: W) {
+        let _ = self[s].insert((t, w));
+    }
+}
+
+impl<W> AddWeightedEdge<W> for Vec<BTreeMap<usize, W>>
+where
+    W: Ord,
+{
+    /// # Panics
+    ///
+    /// Panics if `s` is not in the graph.
+    fn add_weighted_edge(&mut self, s: usize, t: usize, w: W) {
+        let _ = self[s].insert(t, w);
+    }
+}
+
 impl<W> AddWeightedEdge<W> for [Vec<(usize, W)>] {
     /// # Panics
     ///
@@ -141,6 +170,30 @@ where
 impl<W, H> AddWeightedEdge<W> for [HashMap<usize, W, H>]
 where
     H: BuildHasher,
+{
+    /// # Panics
+    ///
+    /// Panics if `s` is not in the graph.
+    fn add_weighted_edge(&mut self, s: usize, t: usize, w: W) {
+        let _ = self[s].insert(t, w);
+    }
+}
+
+impl<W> AddWeightedEdge<W> for [BTreeSet<(usize, W)>]
+where
+    W: Ord,
+{
+    /// # Panics
+    ///
+    /// Panics if `s` is not in the graph.
+    fn add_weighted_edge(&mut self, s: usize, t: usize, w: W) {
+        let _ = self[s].insert((t, w));
+    }
+}
+
+impl<W> AddWeightedEdge<W> for [BTreeMap<usize, W>]
+where
+    W: Ord,
 {
     /// # Panics
     ///
@@ -184,6 +237,30 @@ where
     }
 }
 
+impl<const V: usize, W> AddWeightedEdge<W> for [BTreeSet<(usize, W)>; V]
+where
+    W: Ord,
+{
+    /// # Panics
+    ///
+    /// Panics if `s` is not in the graph.
+    fn add_weighted_edge(&mut self, s: usize, t: usize, w: W) {
+        let _ = self[s].insert((t, w));
+    }
+}
+
+impl<const V: usize, W> AddWeightedEdge<W> for [BTreeMap<usize, W>; V]
+where
+    W: Ord,
+{
+    /// # Panics
+    ///
+    /// Panics if `s` is not in the graph.
+    fn add_weighted_edge(&mut self, s: usize, t: usize, w: W) {
+        let _ = self[s].insert(t, w);
+    }
+}
+
 impl<W, H> AddWeightedEdge<W> for HashMap<usize, Vec<(usize, W)>, H>
 where
     H: BuildHasher,
@@ -207,10 +284,41 @@ where
 impl<W, H> AddWeightedEdge<W> for HashMap<usize, HashMap<usize, W, H>, H>
 where
     H: BuildHasher,
-    HashMap<usize, W, H>: Default,
 {
     fn add_weighted_edge(&mut self, s: usize, t: usize, w: W) {
-        let _ = self.entry(s).or_default().insert(t, w);
+        let _ = self
+            .get_mut(&s)
+            .expect("s is not in the graph")
+            .insert(t, w);
+    }
+}
+
+impl<W> AddWeightedEdge<W> for BTreeMap<usize, Vec<(usize, W)>> {
+    fn add_weighted_edge(&mut self, s: usize, t: usize, w: W) {
+        self.get_mut(&s)
+            .expect("s is not in the graph")
+            .push((t, w));
+    }
+}
+
+impl<W> AddWeightedEdge<W> for BTreeMap<usize, BTreeSet<(usize, W)>>
+where
+    W: Ord,
+{
+    fn add_weighted_edge(&mut self, s: usize, t: usize, w: W) {
+        let _ = self
+            .get_mut(&s)
+            .expect("s is not in the graph")
+            .insert((t, w));
+    }
+}
+
+impl<W> AddWeightedEdge<W> for BTreeMap<usize, BTreeMap<usize, W>> {
+    fn add_weighted_edge(&mut self, s: usize, t: usize, w: W) {
+        let _ = self
+            .get_mut(&s)
+            .expect("s is not in the graph")
+            .insert(t, w);
     }
 }
 
@@ -268,7 +376,7 @@ mod tests {
         assert_eq!(
             graph,
             vec![
-                [(1, 2), (2, 1)].iter().copied().collect(),
+                HashSet::from([(1, 2), (2, 1)]),
                 HashSet::new(),
                 HashSet::new()
             ]
@@ -279,8 +387,8 @@ mod tests {
         assert_eq!(
             graph,
             vec![
-                [(1, 2), (2, 1)].iter().copied().collect(),
-                once((2, 4)).collect(),
+                HashSet::from([(1, 2), (2, 1)]),
+                HashSet::from([(2, 4)]),
                 HashSet::new()
             ]
         );
@@ -290,8 +398,8 @@ mod tests {
         assert_eq!(
             graph,
             vec![
-                [(1, 2), (2, 1)].iter().copied().collect(),
-                [(2, 4), (0, -2)].iter().copied().collect(),
+                HashSet::from([(1, 2), (2, 1)]),
+                HashSet::from([(2, 4), (0, -2)]),
                 HashSet::new()
             ]
         );
@@ -301,9 +409,65 @@ mod tests {
         assert_eq!(
             graph,
             vec![
-                [(1, 2), (2, 1)].iter().copied().collect(),
-                [(2, 4), (0, -2)].iter().copied().collect(),
-                once((0, 3)).collect()
+                HashSet::from([(1, 2), (2, 1)]),
+                HashSet::from([(2, 4), (0, -2)]),
+                HashSet::from([(0, 3)])
+            ]
+        );
+    }
+
+    #[test]
+    fn vec_btree_set() {
+        let mut graph = vec![BTreeSet::new(); 3];
+
+        graph.add_weighted_edge(0, 1, 2);
+
+        assert_eq!(
+            graph,
+            vec![once((1, 2)).collect(), BTreeSet::new(), BTreeSet::new()]
+        );
+
+        graph.add_weighted_edge(0, 2, 1);
+
+        assert_eq!(
+            graph,
+            vec![
+                BTreeSet::from([(1, 2), (2, 1)]),
+                BTreeSet::new(),
+                BTreeSet::new()
+            ]
+        );
+
+        graph.add_weighted_edge(1, 2, 4);
+
+        assert_eq!(
+            graph,
+            vec![
+                BTreeSet::from([(1, 2), (2, 1)]),
+                BTreeSet::from([(2, 4)]),
+                BTreeSet::new()
+            ]
+        );
+
+        graph.add_weighted_edge(1, 0, -2);
+
+        assert_eq!(
+            graph,
+            vec![
+                BTreeSet::from([(1, 2), (2, 1)]),
+                BTreeSet::from([(2, 4), (0, -2)]),
+                BTreeSet::new()
+            ]
+        );
+
+        graph.add_weighted_edge(2, 0, 3);
+
+        assert_eq!(
+            graph,
+            vec![
+                BTreeSet::from([(1, 2), (2, 1)]),
+                BTreeSet::from([(2, 4), (0, -2)]),
+                BTreeSet::from([(0, 3)])
             ]
         );
     }
@@ -316,7 +480,7 @@ mod tests {
 
         assert_eq!(
             graph,
-            vec![once((1, 2)).collect(), HashMap::new(), HashMap::new()]
+            vec![HashMap::from([(1, 2)]), HashMap::new(), HashMap::new()]
         );
 
         graph.add_weighted_edge(0, 2, 1);
@@ -335,8 +499,8 @@ mod tests {
         assert_eq!(
             graph,
             vec![
-                [(1, 2), (2, 1)].iter().copied().collect(),
-                once((2, 4)).collect(),
+                HashMap::from([(1, 2), (2, 1)]),
+                HashMap::from([(2, 4)]),
                 HashMap::new()
             ]
         );
@@ -346,8 +510,8 @@ mod tests {
         assert_eq!(
             graph,
             vec![
-                [(1, 2), (2, 1)].iter().copied().collect(),
-                [(2, 4), (0, -2)].iter().copied().collect(),
+                HashMap::from([(1, 2), (2, 1)]),
+                HashMap::from([(2, 4), (0, -2)]),
                 HashMap::new()
             ]
         );
@@ -357,9 +521,65 @@ mod tests {
         assert_eq!(
             graph,
             vec![
-                [(1, 2), (2, 1)].iter().copied().collect(),
-                [(2, 4), (0, -2)].iter().copied().collect(),
-                once((0, 3)).collect()
+                HashMap::from([(1, 2), (2, 1)]),
+                HashMap::from([(2, 4), (0, -2)]),
+                HashMap::from([(0, 3)]),
+            ]
+        );
+    }
+
+    #[test]
+    fn vec_btree_map() {
+        let mut graph = vec![BTreeMap::new(); 3];
+
+        graph.add_weighted_edge(0, 1, 2);
+
+        assert_eq!(
+            graph,
+            vec![BTreeMap::from([(1, 2)]), BTreeMap::new(), BTreeMap::new()]
+        );
+
+        graph.add_weighted_edge(0, 2, 1);
+
+        assert_eq!(
+            graph,
+            vec![
+                [(1, 2), (2, 1)].into_iter().collect(),
+                BTreeMap::new(),
+                BTreeMap::new()
+            ]
+        );
+
+        graph.add_weighted_edge(1, 2, 4);
+
+        assert_eq!(
+            graph,
+            vec![
+                BTreeMap::from([(1, 2), (2, 1)]),
+                BTreeMap::from([(2, 4)]),
+                BTreeMap::new()
+            ]
+        );
+
+        graph.add_weighted_edge(1, 0, -2);
+
+        assert_eq!(
+            graph,
+            vec![
+                BTreeMap::from([(1, 2), (2, 1)]),
+                BTreeMap::from([(2, 4), (0, -2)]),
+                BTreeMap::new()
+            ]
+        );
+
+        graph.add_weighted_edge(2, 0, 3);
+
+        assert_eq!(
+            graph,
+            vec![
+                BTreeMap::from([(1, 2), (2, 1)]),
+                BTreeMap::from([(2, 4), (0, -2)]),
+                BTreeMap::from([(0, 3)]),
             ]
         );
     }
@@ -453,6 +673,63 @@ mod tests {
     }
 
     #[test]
+    fn slice_btree_set() {
+        let graph: &mut [BTreeSet<(usize, i32)>] =
+            &mut [BTreeSet::new(), BTreeSet::new(), BTreeSet::new()];
+
+        graph.add_weighted_edge(0, 1, 2);
+
+        assert_eq!(
+            *graph,
+            [BTreeSet::from([(1, 2)]), BTreeSet::new(), BTreeSet::new()]
+        );
+
+        graph.add_weighted_edge(0, 2, 1);
+
+        assert_eq!(
+            *graph,
+            [
+                BTreeSet::from([(1, 2), (2, 1)]),
+                BTreeSet::new(),
+                BTreeSet::new()
+            ]
+        );
+
+        graph.add_weighted_edge(1, 2, 4);
+
+        assert_eq!(
+            *graph,
+            [
+                BTreeSet::from([(1, 2), (2, 1)]),
+                BTreeSet::from([(2, 4)]),
+                BTreeSet::new()
+            ]
+        );
+
+        graph.add_weighted_edge(1, 0, -2);
+
+        assert_eq!(
+            *graph,
+            [
+                BTreeSet::from([(1, 2), (2, 1)]),
+                BTreeSet::from([(2, 4), (0, -2)]),
+                BTreeSet::new()
+            ]
+        );
+
+        graph.add_weighted_edge(2, 0, 3);
+
+        assert_eq!(
+            *graph,
+            [
+                BTreeSet::from([(1, 2), (2, 1)]),
+                BTreeSet::from([(2, 4), (0, -2)]),
+                BTreeSet::from([(0, 3)])
+            ]
+        );
+    }
+
+    #[test]
     fn slice_hash_map() {
         let graph: &mut [HashMap<usize, i32>] =
             &mut [HashMap::new(), HashMap::new(), HashMap::new()];
@@ -505,6 +782,63 @@ mod tests {
                 HashMap::from([(1, 2), (2, 1)]),
                 HashMap::from([(2, 4), (0, -2)]),
                 HashMap::from([(0, 3)])
+            ]
+        );
+    }
+
+    #[test]
+    fn slice_btree_map() {
+        let graph: &mut [BTreeMap<usize, i32>] =
+            &mut [BTreeMap::new(), BTreeMap::new(), BTreeMap::new()];
+
+        graph.add_weighted_edge(0, 1, 2);
+
+        assert_eq!(
+            *graph,
+            [BTreeMap::from([(1, 2)]), BTreeMap::new(), BTreeMap::new()]
+        );
+
+        graph.add_weighted_edge(0, 2, 1);
+
+        assert_eq!(
+            *graph,
+            [
+                BTreeMap::from([(1, 2), (2, 1)]),
+                BTreeMap::new(),
+                BTreeMap::new()
+            ]
+        );
+
+        graph.add_weighted_edge(1, 2, 4);
+
+        assert_eq!(
+            *graph,
+            [
+                BTreeMap::from([(1, 2), (2, 1)]),
+                BTreeMap::from([(2, 4)]),
+                BTreeMap::new()
+            ]
+        );
+
+        graph.add_weighted_edge(1, 0, -2);
+
+        assert_eq!(
+            *graph,
+            [
+                BTreeMap::from([(1, 2), (2, 1)]),
+                BTreeMap::from([(2, 4), (0, -2)]),
+                BTreeMap::new()
+            ]
+        );
+
+        graph.add_weighted_edge(2, 0, 3);
+
+        assert_eq!(
+            *graph,
+            [
+                BTreeMap::from([(1, 2), (2, 1)]),
+                BTreeMap::from([(2, 4), (0, -2)]),
+                BTreeMap::from([(0, 3)])
             ]
         );
     }
@@ -597,6 +931,62 @@ mod tests {
     }
 
     #[test]
+    fn arr_btree_set() {
+        let mut graph = [BTreeSet::new(), BTreeSet::new(), BTreeSet::new()];
+
+        graph.add_weighted_edge(0, 1, 2);
+
+        assert_eq!(
+            graph,
+            [BTreeSet::from([(1, 2)]), BTreeSet::new(), BTreeSet::new()]
+        );
+
+        graph.add_weighted_edge(0, 2, 1);
+
+        assert_eq!(
+            graph,
+            [
+                BTreeSet::from([(1, 2), (2, 1)]),
+                BTreeSet::new(),
+                BTreeSet::new()
+            ]
+        );
+
+        graph.add_weighted_edge(1, 2, 4);
+
+        assert_eq!(
+            graph,
+            [
+                BTreeSet::from([(1, 2), (2, 1)]),
+                BTreeSet::from([(2, 4)]),
+                BTreeSet::new()
+            ]
+        );
+
+        graph.add_weighted_edge(1, 0, -2);
+
+        assert_eq!(
+            graph,
+            [
+                BTreeSet::from([(1, 2), (2, 1)]),
+                BTreeSet::from([(2, 4), (0, -2)]),
+                BTreeSet::new()
+            ]
+        );
+
+        graph.add_weighted_edge(2, 0, 3);
+
+        assert_eq!(
+            graph,
+            [
+                BTreeSet::from([(1, 2), (2, 1)]),
+                BTreeSet::from([(2, 4), (0, -2)]),
+                BTreeSet::from([(0, 3)])
+            ]
+        );
+    }
+
+    #[test]
     fn arr_hash_map() {
         let mut graph = [HashMap::new(), HashMap::new(), HashMap::new()];
 
@@ -648,6 +1038,62 @@ mod tests {
                 HashMap::from([(1, 2), (2, 1)]),
                 HashMap::from([(2, 4), (0, -2)]),
                 HashMap::from([(0, 3)])
+            ]
+        );
+    }
+
+    #[test]
+    fn arr_btree_map() {
+        let mut graph = [BTreeMap::new(), BTreeMap::new(), BTreeMap::new()];
+
+        graph.add_weighted_edge(0, 1, 2);
+
+        assert_eq!(
+            graph,
+            [BTreeMap::from([(1, 2)]), BTreeMap::new(), BTreeMap::new()]
+        );
+
+        graph.add_weighted_edge(0, 2, 1);
+
+        assert_eq!(
+            graph,
+            [
+                BTreeMap::from([(1, 2), (2, 1)]),
+                BTreeMap::new(),
+                BTreeMap::new()
+            ]
+        );
+
+        graph.add_weighted_edge(1, 2, 4);
+
+        assert_eq!(
+            graph,
+            [
+                BTreeMap::from([(1, 2), (2, 1)]),
+                BTreeMap::from([(2, 4)]),
+                BTreeMap::new()
+            ]
+        );
+
+        graph.add_weighted_edge(1, 0, -2);
+
+        assert_eq!(
+            graph,
+            [
+                BTreeMap::from([(1, 2), (2, 1)]),
+                BTreeMap::from([(2, 4), (0, -2)]),
+                BTreeMap::new()
+            ]
+        );
+
+        graph.add_weighted_edge(2, 0, 3);
+
+        assert_eq!(
+            graph,
+            [
+                BTreeMap::from([(1, 2), (2, 1)]),
+                BTreeMap::from([(2, 4), (0, -2)]),
+                BTreeMap::from([(0, 3)])
             ]
         );
     }
@@ -736,15 +1182,33 @@ mod tests {
 
     #[test]
     fn hash_map_hash_map() {
-        let mut graph = HashMap::new();
+        let mut graph = HashMap::from([
+            (0, HashMap::new()),
+            (1, HashMap::new()),
+            (2, HashMap::new()),
+        ]);
 
         graph.add_weighted_edge(0, 1, 2);
 
-        assert_eq!(graph, HashMap::from([(0, HashMap::from([(1, 2)]))]));
+        assert_eq!(
+            graph,
+            HashMap::from([
+                (0, HashMap::from([(1, 2)])),
+                (1, HashMap::new()),
+                (2, HashMap::new())
+            ])
+        );
 
         graph.add_weighted_edge(0, 2, 1);
 
-        assert_eq!(graph, HashMap::from([(0, HashMap::from([(1, 2), (2, 1)]))]));
+        assert_eq!(
+            graph,
+            HashMap::from([
+                (0, HashMap::from([(1, 2), (2, 1)])),
+                (1, HashMap::new()),
+                (2, HashMap::new())
+            ])
+        );
 
         graph.add_weighted_edge(1, 2, 4);
 
@@ -753,6 +1217,7 @@ mod tests {
             HashMap::from([
                 (0, HashMap::from([(1, 2), (2, 1)])),
                 (1, HashMap::from([(2, 4)])),
+                (2, HashMap::new())
             ])
         );
 
@@ -763,6 +1228,7 @@ mod tests {
             HashMap::from([
                 (0, HashMap::from([(1, 2), (2, 1)])),
                 (1, HashMap::from([(2, 4), (0, -2)])),
+                (2, HashMap::new())
             ])
         );
 
@@ -774,6 +1240,186 @@ mod tests {
                 (0, HashMap::from([(1, 2), (2, 1)])),
                 (1, HashMap::from([(2, 4), (0, -2)])),
                 (2, HashMap::from([(0, 3)])),
+            ])
+        );
+    }
+
+    #[test]
+    fn btree_map_vec() {
+        let mut graph = BTreeMap::from([(0, Vec::new()), (1, Vec::new()), (2, Vec::new())]);
+
+        graph.add_weighted_edge(0, 1, 2);
+
+        assert_eq!(
+            graph,
+            BTreeMap::from([(0, vec![(1, 2)]), (1, Vec::new()), (2, Vec::new())])
+        );
+
+        graph.add_weighted_edge(0, 2, 1);
+
+        assert_eq!(
+            graph,
+            BTreeMap::from([(0, vec![(1, 2), (2, 1)]), (1, Vec::new()), (2, Vec::new())])
+        );
+
+        graph.add_weighted_edge(1, 2, 4);
+
+        assert_eq!(
+            graph,
+            BTreeMap::from([
+                (0, vec![(1, 2), (2, 1)]),
+                (1, vec![(2, 4)]),
+                (2, Vec::new())
+            ])
+        );
+
+        graph.add_weighted_edge(1, 0, -2);
+
+        assert_eq!(
+            graph,
+            BTreeMap::from([
+                (0, vec![(1, 2), (2, 1)]),
+                (1, vec![(2, 4), (0, -2)]),
+                (2, Vec::new())
+            ])
+        );
+
+        graph.add_weighted_edge(2, 0, 3);
+
+        assert_eq!(
+            graph,
+            BTreeMap::from([
+                (0, vec![(1, 2), (2, 1)]),
+                (1, vec![(2, 4), (0, -2)]),
+                (2, vec![(0, 3)])
+            ])
+        );
+    }
+
+    #[test]
+    fn btree_map_btree_set() {
+        let mut graph = BTreeMap::from([
+            (0, BTreeSet::new()),
+            (1, BTreeSet::new()),
+            (2, BTreeSet::new()),
+        ]);
+
+        graph.add_weighted_edge(0, 1, 2);
+
+        assert_eq!(
+            graph,
+            BTreeMap::from([
+                (0, BTreeSet::from([(1, 2)])),
+                (1, BTreeSet::new()),
+                (2, BTreeSet::new())
+            ])
+        );
+
+        graph.add_weighted_edge(0, 2, 1);
+
+        assert_eq!(
+            graph,
+            BTreeMap::from([
+                (0, BTreeSet::from([(1, 2), (2, 1)])),
+                (1, BTreeSet::new()),
+                (2, BTreeSet::new())
+            ])
+        );
+
+        graph.add_weighted_edge(1, 2, 4);
+
+        assert_eq!(
+            graph,
+            BTreeMap::from([
+                (0, BTreeSet::from([(1, 2), (2, 1)])),
+                (1, BTreeSet::from([(2, 4)])),
+                (2, BTreeSet::new())
+            ])
+        );
+
+        graph.add_weighted_edge(1, 0, -2);
+
+        assert_eq!(
+            graph,
+            BTreeMap::from([
+                (0, BTreeSet::from([(1, 2), (2, 1)])),
+                (1, BTreeSet::from([(2, 4), (0, -2)])),
+                (2, BTreeSet::new())
+            ])
+        );
+
+        graph.add_weighted_edge(2, 0, 3);
+
+        assert_eq!(
+            graph,
+            BTreeMap::from([
+                (0, BTreeSet::from([(1, 2), (2, 1)])),
+                (1, BTreeSet::from([(2, 4), (0, -2)])),
+                (2, BTreeSet::from([(0, 3)])),
+            ])
+        );
+    }
+
+    #[test]
+    fn btree_map_btree_map() {
+        let mut graph = BTreeMap::from([
+            (0, BTreeMap::new()),
+            (1, BTreeMap::new()),
+            (2, BTreeMap::new()),
+        ]);
+
+        graph.add_weighted_edge(0, 1, 2);
+
+        assert_eq!(
+            graph,
+            BTreeMap::from([
+                (0, BTreeMap::from([(1, 2)])),
+                (1, BTreeMap::new()),
+                (2, BTreeMap::new())
+            ])
+        );
+
+        graph.add_weighted_edge(0, 2, 1);
+
+        assert_eq!(
+            graph,
+            BTreeMap::from([
+                (0, BTreeMap::from([(1, 2), (2, 1)])),
+                (1, BTreeMap::new()),
+                (2, BTreeMap::new())
+            ])
+        );
+
+        graph.add_weighted_edge(1, 2, 4);
+
+        assert_eq!(
+            graph,
+            BTreeMap::from([
+                (0, BTreeMap::from([(1, 2), (2, 1)])),
+                (1, BTreeMap::from([(2, 4)])),
+                (2, BTreeMap::new())
+            ])
+        );
+
+        graph.add_weighted_edge(1, 0, -2);
+
+        assert_eq!(
+            graph,
+            BTreeMap::from([
+                (0, BTreeMap::from([(1, 2), (2, 1)])),
+                (1, BTreeMap::from([(2, 4), (0, -2)])),
+                (2, BTreeMap::new())
+            ])
+        );
+
+        graph.add_weighted_edge(2, 0, 3);
+
+        assert_eq!(
+            graph,
+            BTreeMap::from([
+                (0, BTreeMap::from([(1, 2), (2, 1)])),
+                (1, BTreeMap::from([(2, 4), (0, -2)])),
+                (2, BTreeMap::from([(0, 3)])),
             ])
         );
     }